fix examples and depencencies
