+++ javascript/modules/client/examples/00-client/05-get-daos.ts
@@ -12,7 +12,7 @@ import {
   IDaoQueryParams,
   SortDirection,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "./00-context";
 
 const context: Context = new Context(contextParams);
 const client: Client = new Client(context);
