+++ javascript/modules/client/examples/01-client-erc20/00-installation.ts
@@ -18,7 +18,7 @@ import {
   ICreateParams,
   IErc20PluginInstall,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 const context: Context = new Context(contextParams);
 const client: Client = new Client(context);
