+++ javascript/modules/client/examples/01-client-erc20/01-context.ts
@@ -3,7 +3,7 @@
 */
 import { Context, ContextPlugin } from "@aragon/sdk-client";
 import { Wallet } from "@ethersproject/wallet";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 const context = new Context(contextParams);
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
