+++ javascript/modules/client/examples/01-client-erc20/02-client.ts
@@ -2,7 +2,7 @@
 ### Create an ERC20 client
 */
 import { ClientErc20, Context, ContextPlugin } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 const context = new Context(contextParams);
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
