+++ javascript/modules/client/examples/01-client-erc20/03-create-proposal.ts
@@ -9,7 +9,7 @@ import {
   ProposalCreationSteps,
   VoteValues,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 // Create a simple context
 const context: Context = new Context(contextParams);
