+++ javascript/modules/client/examples/01-client-erc20/09-get-proposals.ts
@@ -13,7 +13,7 @@ import {
   SortDirection,
   ProposalStatus,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 // Create a simple context
 const context: Context = new Context(contextParams);
