+++ javascript/modules/client/examples/02-client-addressList/00-installation.ts
@@ -11,7 +11,7 @@ import {
   IAddressListPluginInstall,
   ICreateParams,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 const context: Context = new Context(contextParams);
 const client: Client = new Client(context);
