+++ javascript/modules/client/examples/02-client-addressList/04-create-proposal-with-actions.ts
@@ -10,7 +10,7 @@ import {
   ProposalCreationSteps,
   VoteValues,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 // Create a simple context
 const context: Context = new Context(contextParams);
