+++ javascript/modules/client/examples/02-client-addressList/05-vote-proposal.ts
@@ -9,7 +9,7 @@ import {
   VoteProposalStep,
   VoteValues,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 // Create a simple context
 const context: Context = new Context(contextParams);
