+++ javascript/modules/client/examples/02-client-addressList/06-can-vote.ts
@@ -7,7 +7,7 @@ import {
   ContextPlugin,
   ICanVoteParams,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 // Create a simple context
 const context: Context = new Context(contextParams);
