+++ javascript/modules/client/examples/02-client-addressList/07-get-members.ts
@@ -2,7 +2,7 @@
 ### Loading the list of members (address list plugin)
 */
 import { ClientAddressList, Context, ContextPlugin } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 // Create a simple context
 const context: Context = new Context(contextParams);
