+++ javascript/modules/client/examples/02-client-addressList/09-get-proposals.ts
@@ -11,7 +11,7 @@ import {
   ProposalStatus,
   SortDirection,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 // Create a simple context
 const context: Context = new Context(contextParams);
