+++ javascript/modules/client/examples/03-encoders/03-freeze-permissions.ts
@@ -7,7 +7,7 @@ import {
   IFreezePermissionParams,
   Permissions,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 const context: Context = new Context(contextParams);
 const client: Client = new Client(context);
