+++ javascript/modules/client/examples/03-encoders/04-withdraw.ts
@@ -2,7 +2,7 @@
 ### Withdrawals
 */
 import { Client, Context, IWithdrawParams } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 const context: Context = new Context(contextParams);
 const client: Client = new Client(context);
