+++ javascript/modules/client/examples/03-encoders/07-update-plugin-settings-erc-20.ts
@@ -7,7 +7,7 @@ import {
   ContextPlugin,
   IPluginSettings,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 const context: Context = new Context(contextParams);
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
