+++ javascript/modules/client/examples/03-encoders/08-mint-token.ts
@@ -7,7 +7,7 @@ import {
   ContextPlugin,
   IMintTokenParams,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 const context: Context = new Context(contextParams);
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
