+++ javascript/modules/client/examples/03-encoders/09-add-members.ts
@@ -2,7 +2,7 @@
 ### Add Members (AddressList)
 */
 import { ClientAddressList, Context, ContextPlugin } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 const context: Context = new Context(contextParams);
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
