+++ javascript/modules/client/examples/03-encoders/10-remove-members.ts
@@ -6,7 +6,7 @@ import {
   Context,
   ContextPlugin,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 const context: Context = new Context(contextParams);
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
