+++ javascript/modules/client/examples/04-decoders/02-revoke-permission.ts
@@ -6,7 +6,7 @@ import {
   Context,
   IRevokePermissionDecodedParams,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 const context: Context = new Context(contextParams);
 const client: Client = new Client(context);
