+++ javascript/modules/client/examples/04-decoders/05-update-metadata.ts
@@ -4,7 +4,7 @@
 Decode an update metadata action and expect the metadata
 */
 import { Client, Context, IMetadata } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 const context: Context = new Context(contextParams);
 const client: Client = new Client(context);
 const data: Uint8Array = new Uint8Array([12, 56]);
