+++ javascript/modules/client/examples/04-decoders/07-update-plugin-settings-address-list.ts
@@ -7,7 +7,7 @@ import {
   ContextPlugin,
   IPluginSettings,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 const context: Context = new Context(contextParams);
 // Create a plugin context from the simple context
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
