+++ javascript/modules/client/examples/04-decoders/10-find-interface-address-list.ts
@@ -6,7 +6,7 @@ import {
   Context,
   ContextPlugin,
 } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 const context: Context = new Context(contextParams);
 // Create a plugin context from the simple context
