+++ javascript/modules/client/examples/04-decoders/11-find-interface-erc-20.ts
@@ -2,7 +2,7 @@
 ### Get Function Parameters from an encoded action (ERC-20)
 */
 import { ClientErc20, Context, ContextPlugin } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 
 const context: Context = new Context(contextParams);
 // Create a plugin context from the simple context
