+++ javascript/modules/client/examples/04-decoders/14-remove-members.ts
@@ -2,7 +2,7 @@
 ### Decode Remove Members Action (Address List)
 */
 import { ClientAddressList, Context, ContextPlugin } from "@aragon/sdk-client";
-import { contextParams } from "../context";
+import { contextParams } from "../00-client/00-context";
 const context: Context = new Context(contextParams);
 // Create a plugin context from the simple context
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
