+++ javascript/modules/client/package.json
@@ -49,6 +49,7 @@
   "devDependencies": {
     "@size-limit/preset-small-lib": "^7.0.8",
     "@types/jest": "^29.0.3",
+    "glob": "^8.0.3",
     "husky": "^7.0.4",
     "size-limit": "^7.0.8",
     "tsdx": "^0.14.1",
